Disable VIGRANumPy by default with GNU c++ < 4.4 on MacOSX
This developer's machine is basically stuck with GNU C++ 4.2 as that is
the last XCode version available for MacOSX 10.6 (and the machine is not
upgradeable).

Rather than leaving other developers with the same problem puzzled about a
weird and long compile error, let's be nice and just disable VIGRANumPy in
that case.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>