travis: Switch to Python 3.4 for testing on Mac.
Recent versions of pip and setuptools have dropped support for Python
versions before 3.3. This makes it hard to install them into using their
installation script. What `pyenv` does is use the most recent copy of
this script, which is causing this version to fail for us. Instead let's
test on a newer version of Python, which is pretty more stable (3.4). We
plan on testing this for Linux, as well.