travis: Install Boost using verbose mode.
Travis kills any build step taking more than 10 min and not producing
output to the consolve. Trying to use `travis_wait` here seems to not be
working correctly. Here we try simply to increase the verbosity so as to
ensure Travis won't kill the step even if it takes awhile.