Revert "travis: Try to disable python with Boost.Python."
This reverts commit 0311276f837ba286d4c455d0dc984ce221c33479.

Appears this just results in no Python build at all. We will have to
hope for the best when mixing the Python and pyenv's Python libraries
together. If we are still running into issues, we could try to fix the
linkage after the fact. However, there are no ways to do this through
brew that I can see so it will have to be done manually.