appveyor: Add a build matrix with Python 2 & 3.
In this patch, we install different versions of Python and miniconda. We
try to keep miniconda with the same version that we will use in our test
environment. We do this because if we have to start using `conda-build`
for some reason the version of the built packages will be determined by
the root environment and we want to make sure they match up with the
environment we wish to install them into.

Also, our initial version had some issue with VC++ versions. This has
been fixed thanks to some feedback from Ulli.