travis: Uninstall only unneeded packages from brew.
Travis ships a bunch of packages that we don't need, may be outdated, or
are somehow stopping are ability to update regardless of whether we use
them. Previously we solved this problem by simply removing all of them.
However, this is making the install step take too long and possibly
causing builds to be terminated as a consequence. This adds a check to
list all of our dependencies and all their dependencies and only removes
packages that we will not use or require in some fashion from the
system. For this to work correctly, we need to make sure all
recipes/packages that we are looking for are present. So, we update the
brew recipes (doesn't install anything) and tap the relevant additional
recipes that we need.