travis: Update pip and setuptools on all supported platforms.
This is recommended by homebrew as its bottles may not contain the
latest versions of these tools. The same is generally true of the Travis
CI's Linux VMs' copies of these tools. So, go ahead and update them for
all Travis CI VMs. Handle Python 3.2 separately as it is only supported
by certain versions of setuptools and pip.