travis: Skip updating brew.
This is the only step that really has an effect over how long the builds
take at this point. By not updating, it avoids both spending the time
updating and the much more costly time spent replacing outdated
packages. Instead it will now easily be able to use all outdated
packages as they are update to according to the current state of brew as
provided to us by Travis CI.