Set VIGRA_CONFIGURATION always to CMAKE_CFG_INTDIR
Always set `VIGRA_CONFIGURATION` to `CMAKE_CFG_INTDIR`. This amounts to
it being `Release`, `Debug`, or `.`. The latter occurring in the NMake
case. The result being that we always check a valid directory. To make
sure the `run_test.bat` script is reasonable. Make sure the `.` case is
explained to the user as the `Release` case as it would have been
before.