Receive `const char *` from `PyUnicode_AsUTF8`
In Python 3.7, `PyUnicode_AsUTF8` was changed to return a `const char *`
instead of a `char *`. This broke VIGRA as we were accepting a `char *`
in this case instead. Fortunately we do not need it to be mutable for
our use case. So just type the variable storing the result from
`PyUnicode_AsUTF8` as a `const char *`. Should still work on older
Python 3 versions that return `char *` as well.

ref: https://bugs.python.org/issue28769