use tuples for numpy indexing
`numpy` currently warns with the following warning when not using tuples for
indexing:

```
vigranumpy/vigra/arraytypes.py:1263: FutureWarning: Using a non-tuple sequence
for multidimensional indexing is deprecated; use `arr[tuple(seq)]` instead of `arr[seq]`.
In the future this will be interpreted as an array index, `arr[np.array(seq)]`, which will
result either in an error or a different result.
```

fixed by simply converting the index to a tuple in relevant places.

fixes #474