build: update cmake from 2.6 to 3.10
CMake 2.6 is really ancient and deprecated: newer CMake will stop
supporting it. Updating also allows to use modern CMake tecniques.

This also makes it easier for other CMake projects to use VIGRA
(I stumbled on this while trying to build Hugin).

To give an idea, CMake is currently at 3.19. CMake 3.10 is what is
shipped with Ubuntu 18.04, which has been released in April 2018.

Tested on macOS 10.15 Catalina.